// src/app/api/api/financials.service.ts
export interface ApiFinancialsSearchApiFinancialsSearchGetRequestParams {
    /** Field to sort by (e.g., market_cap, pe, dividend_yield, revenue_growth_yoy) */
    sortBy?: string;
    sortDir?: SortDir;
    page?: number;
    pageSize?: number;
    sector?: GICSSector | null;
    industry?: string | null;
    minMcap?: number | null;
    maxMcap?: number | null;
    minPe?: number | null;
    maxPe?: number | null;
    minPb?: number | null;
    maxPb?: number | null;
    minPs?: number | null;
    maxPs?: number | null;
    minDividendYield?: number | null;
    maxDividendYield?: number | null;
    maxPayoutRatio?: number | null;
    maxNetDebtToEbitda?: number | null;
    minRevenueGrowthYoy?: number | null;
    minEpsGrowthYoy?: number | null;
    frequency?: number | null;
}